+++ R/layouts.R
@@ -22,6 +22,8 @@
 #' * `topenv`: the name of the top environment from which the parent call was called
 #'   (eg R package name or `GlobalEnv`)
 #' * `call`: parent call (if any) calling the logging function
+#' * `location`: A list with element `path` and `line` giving the location of the
+#'   log call
 #' * `fn`: function's (if any) name calling the logging function
 #'
 #' @param log_level log level as per [log_levels()]
@@ -35,9 +37,9 @@ get_logger_meta_variables <- function(log_level = NULL,
                                       namespace = NA_character_,
                                       .logcall = sys.call(),
                                       .topcall = sys.call(-1),
-                                      .topenv = parent.frame()) {
+                                      .topenv = parent.frame(),
+                                      .timestamp = Sys.time()) {
   sysinfo <- Sys.info()
-  timestamp <- Sys.time()
 
   list(
     ns = namespace,
@@ -45,7 +47,8 @@ get_logger_meta_variables <- function(log_level = NULL,
     topenv = top_env_name(.topenv),
     fn = deparse_to_one_line(.topcall[[1]]),
     call = deparse_to_one_line(.topcall),
-    time = timestamp,
+    location = log_call_location(.logcall),
+    time = .timestamp,
     levelr = log_level,
     level = attr(log_level, "level"),
     pid = Sys.getpid(),
@@ -71,10 +74,10 @@ get_logger_meta_variables <- function(log_level = NULL,
 
 #' Generate log layout function using common variables available via glue syntax
 #'
-#' `format` is passed to `glue` with access to the below variables:
+#' `format` is passed to [glue::glue()] with access to the below variables:
 #' \itemize{ \item msg: the actual log message \item further variables
 #' set by [get_logger_meta_variables()] }
-#' @param format `glue`-flavored layout of the message using the above
+#' @param format [glue::glue()]-flavored layout of the message using the above
 #'   variables
 #' @return function taking `level` and `msg` arguments - keeping the
 #'   original call creating the generator in the `generator` attribute
@@ -101,7 +104,8 @@ layout_glue_generator <- function(format = '{level} [{format(time, "%Y-%m-%d %H:
                      namespace = NA_character_,
                      .logcall = sys.call(),
                      .topcall = sys.call(-1),
-                     .topenv = parent.frame()) {
+                     .topenv = parent.frame(),
+                     .timestamp = Sys.time()) {
     fail_on_missing_package("glue")
     if (!inherits(level, "loglevel")) {
       stop("Invalid log level, see ?log_levels")
@@ -113,6 +117,7 @@ layout_glue_generator <- function(format = '{level} [{format(time, "%Y-%m-%d %H:
       .logcall = .logcall,
       .topcall = .topcall,
       .topenv = .topenv,
+      .timestamp = .timestamp,
       parent = environment()
     )
     glue::glue(format, .envir = meta)
@@ -132,7 +137,8 @@ layout_blank <- function(level,
                          namespace = NA_character_,
                          .logcall = sys.call(),
                          .topcall = sys.call(-1),
-                         .topenv = parent.frame()) {
+                         .topenv = parent.frame(),
+                         .timestamp = Sys.time()) {
   msg
 }
 attr(layout_blank, "generator") <- quote(layout_blank())
@@ -149,8 +155,9 @@ layout_simple <- function(level,
                           namespace = NA_character_,
                           .logcall = sys.call(),
                           .topcall = sys.call(-1),
-                          .topenv = parent.frame()) {
-  paste0(attr(level, "level"), " [", format(Sys.time(), "%Y-%m-%d %H:%M:%S"), "] ", msg)
+                          .topenv = parent.frame(),
+                          .timestamp = Sys.time()) {
+  paste0(attr(level, "level"), " [", format(.timestamp, "%Y-%m-%d %H:%M:%S"), "] ", msg)
 }
 attr(layout_simple, "generator") <- quote(layout_simple())
 
@@ -176,16 +183,18 @@ layout_logging <- function(level,
                            namespace = NA_character_,
                            .logcall = sys.call(),
                            .topcall = sys.call(-1),
-                           .topenv = parent.frame()) {
+                           .topenv = parent.frame(),
+                           .timestamp = Sys.time()) {
   meta <- logger_meta_env(
     log_level = level,
     namespace = namespace,
     .logcall = .logcall,
     .topcall = .topcall,
-    .topenv = .topenv
+    .topenv = .topenv,
+    .timestamp = .timestamp
   )
   paste0(
-    format(Sys.time(), "%Y-%m-%d %H:%M:%S"), " ",
+    format(meta$time, "%Y-%m-%d %H:%M:%S"), " ",
     attr(level, "level"), ":",
     ifelse(meta$ns == "global", "", meta$ns), ":",
     msg
@@ -193,7 +202,7 @@ layout_logging <- function(level,
 }
 attr(layout_logging, "generator") <- quote(layout_logging())
 
-#' Format a log message with `glue`
+#' Format a log message with [glue::glue()]
 #'
 #' By default, this layout includes the log level of the log record as
 #' per [log_levels()], the current timestamp and the actual log
@@ -207,7 +216,7 @@ attr(layout_logging, "generator") <- quote(layout_logging())
 layout_glue <- layout_glue_generator()
 attr(layout_glue, "generator") <- quote(layout_glue())
 
-#' Format a log message with `glue` and ANSI escape codes to add colors
+#' Format a log message with [glue::glue()] and ANSI escape codes to add colors
 #'
 #' Colour log levels based on their severity. Log levels are coloured
 #' with [colorize_by_log_level()] and the messages are coloured with
@@ -264,7 +273,8 @@ layout_json <- function(fields = default_fields()) {
                      namespace = NA_character_,
                      .logcall = sys.call(),
                      .topcall = sys.call(-1),
-                     .topenv = parent.frame()) {
+                     .topenv = parent.frame(),
+                     .timestamp = Sys.time()) {
     fail_on_missing_package("jsonlite")
 
     meta <- logger_meta_env(
@@ -272,7 +282,8 @@ layout_json <- function(fields = default_fields()) {
       namespace = namespace,
       .logcall = .logcall,
       .topcall = .topcall,
-      .topenv = .topenv
+      .topenv = .topenv,
+      .timestamp = .timestamp
     )
     json <- mget(fields, meta)
     sapply(msg, function(msg) jsonlite::toJSON(c(json, list(msg = msg)), auto_unbox = TRUE))
@@ -283,7 +294,7 @@ layout_json <- function(fields = default_fields()) {
 #' Generate log layout function rendering JSON after merging meta
 #' fields with parsed list from JSON message
 #' @param fields character vector of field names to be included in the
-#'   JSON
+#'   JSON. If named, the names will be used as field names in the JSON.
 #' @export
 #' @note This functionality depends on the \pkg{jsonlite} package.
 #' @family log_layouts
@@ -297,7 +308,16 @@ layout_json <- function(fields = default_fields()) {
 #'
 #' log_layout(layout_json_parser(fields = c("time", "node")))
 #' log_info(cars = row.names(mtcars), species = unique(iris$Species))
+#'
+#' log_layout(layout_json_parser(fields = c(timestamp = "time", "node")))
+#' log_info(
+#'   message = paste(
+#'     "Compared to the previous example,
+#'     the 'time' field is renamed to 'timestamp'"
+#'   )
+#' )
 #' \dontshow{logger:::namespaces_set(old)}
+#' @importFrom stats setNames
 layout_json_parser <- function(fields = default_fields()) {
   force(fields)
 
@@ -306,7 +326,8 @@ layout_json_parser <- function(fields = default_fields()) {
                      namespace = NA_character_,
                      .logcall = sys.call(),
                      .topcall = sys.call(-1),
-                     .topenv = parent.frame()) {
+                     .topenv = parent.frame(),
+                     .timestamp = Sys.time()) {
     fail_on_missing_package("jsonlite")
 
     meta <- logger_meta_env(
@@ -314,9 +335,16 @@ layout_json_parser <- function(fields = default_fields()) {
       namespace = namespace,
       .logcall = .logcall,
       .topcall = .topcall,
-      .topenv = .topenv
+      .topenv = .topenv,
+      .timestamp = .timestamp
     )
     meta <- mget(fields, meta)
+    field_names <- names(fields)
+    if (!is.null(field_names)) {
+      norename <- field_names == ""
+      field_names[norename] <- fields[norename]
+      meta <- setNames(meta, field_names)
+    }
     msg <- jsonlite::fromJSON(msg)
 
     jsonlite::toJSON(c(meta, msg), auto_unbox = TRUE, null = "null")
@@ -329,6 +357,48 @@ default_fields <- function() {
     "os_name", "os_release", "os_version", "pid", "user"
   )
 }
+
+#' Format a log record for github actions
+#'
+#' GitHub Actions can recognise specially formatted output and make these
+#' prominent in the output. The `layout_gha()` layout will ensure the correct
+#' formatting when running in GitHub Actions.
+#'
+#' @note
+#' GitHub Actions only recognise the log levels `error`, `warning`, `notice`,
+#' and `debug`. Because of this, `FATAL` and `ERROR` are coerced to `error`,
+#' `SUCCESS` and `INFO` are coerced to `notice`, and `DEBUG` and `TRACE` are
+#' coerced to `debug` (`WARN` maps directly to `warning`).
+#'
+#' @inheritParams layout_simple
+#' @return character vector
+#' @export
+#' @family log_layouts
+#'
+layout_gha <- structure(
+  function(level,
+           msg,
+           namespace = NA_character_,
+           .logcall = sys.call(),
+           .topcall = sys.call(-1),
+           .topenv = parent.frame(),
+           .timestamp = Sys.time()) {
+    level <- attr(level, "level")
+    gha_level <- switch(level,
+      FATAL = ,
+      ERROR = "error",
+      WARN = "warning",
+      SUCCESS = ,
+      INFO = "notice",
+      DEBUG = ,
+      TRACE = "debug",
+      "notice" # Defaults to notice if for some reason another level gets used
+    )
+    title <- if (gha_level == "debug") "" else paste0(" title=", level)
+    paste0("::", gha_level, title, "::", msg)
+  },
+  generator = quote(layout_gha())
+)
 
 # nocov start
 #' Format a log record for syslognet
@@ -346,7 +416,8 @@ layout_syslognet <- structure(
            namespace = NA_character_,
            .logcall = sys.call(),
            .topcall = sys.call(-1),
-           .topenv = parent.frame()) {
+           .topenv = parent.frame(),
+           .timestamp = Sys.time()) {
     ret <- paste(attr(level, "level"), msg)
     attr(ret, "severity") <- switch(
       attr(level, "level", exact = TRUE),
@@ -363,4 +434,3 @@ layout_syslognet <- structure(
   generator = quote(layout_syslognet())
 )
 # nocov end
-